Refactor: Option, OptionCategory id Long에서 Integer로 수정

+++ src/main/java/com/zerobase/CafeBom/option/repository/OptionCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OptionCategoryRepository extends JpaRepository<OptionCategory, Long> {
+public interface OptionCategoryRepository extends JpaRepository<OptionCategory, Integer> {
 }
